build: silence all compile warnings
Project sources now build with zero warnings under -Wall -Wextra.

- Makefile: pull the PipeWire/SPA headers in as system headers (-isystem
  instead of -I) so two warnings from inside /usr/include/spa-0.2
  (spa_param_info missing 'user' initializer, compare.h unused 'size')
  are not reported; our own sources keep the full -Wall -Wextra
  treatment.  Link test-pw-dmabuf-consumer against libpipewire-0.3
  (it was failing to link: pw_* symbols undefined).
- src/camhal_backend.cpp: drop redundant (uint32_t) casts on the
  stream_t width/height comparisons (stream_t fields are already int),
  clearing six -Wsign-compare warnings.
- test/test-pw-dmabuf-direct.c, test/test-pw-dmabuf-consumer.c: print
  spa_data.fd with PRIi64 (it is int64_t), drop an unused function and
  unused parameters.

+++ src/camhal_backend.cpp
@@ -831,8 +831,8 @@ int camhal_backend_configure(struct camhal_backend *b,
 		get_supported_stream_config(cinfo.capability, &configs);
 		for (size_t i = 0; i < configs.size(); i++) {
 			if ((uint32_t)configs[i].format == format &&
-			    configs[i].width == (uint32_t)width &&
-			    configs[i].height == (uint32_t)height) {
+			    configs[i].width == width &&
+			    configs[i].height == height) {
 				stream = configs[i];
 				found = 1;
 				break;
@@ -1167,8 +1167,8 @@ int camhal_backend_configure_external(struct camhal_backend *b,
 		get_supported_stream_config(cinfo.capability, &configs);
 		for (size_t i = 0; i < configs.size(); i++) {
 			if ((uint32_t)configs[i].format == format &&
-			    configs[i].width == (uint32_t)width &&
-			    configs[i].height == (uint32_t)height) {
+			    configs[i].width == width &&
+			    configs[i].height == height) {
 				stream = configs[i];
 				found = 1;
 				break;
@@ -1271,8 +1271,8 @@ int camhal_backend_configure_dmabuf(struct camhal_backend *b,
 		get_supported_stream_config(cinfo.capability, &configs);
 		for (size_t i = 0; i < configs.size(); i++) {
 			if ((uint32_t)configs[i].format == format &&
-			    configs[i].width == (uint32_t)width &&
-			    configs[i].height == (uint32_t)height) {
+			    configs[i].width == width &&
+			    configs[i].height == height) {
 				stream = configs[i];
 				found = 1;
 				break;
